Created an equals method in AbstractVehicle and early version of displayVehicleInformation
I was unsure it displayVehicleInformation needed to utilize a equals method, so I created one for AbstractVehicle. It compares all of the fields. displayVehicleInformation looks for an object that is identical to the actual object (without using an equals method) currently. Also added some getters/setters in AbstractVehicle (specifically for Make and GasTankCapacity)