implemented saveVehicleList (and helper methods), minor bugfix
SaveVehicleList was implemented in VehicleManagerSingleton class. It has a helper method to generates a string to write to the file called fileCompatibleStringGenerator. Also implemented getClassString in the AbstractVehicle, as well as all the non-abstract vehicles. getClassString is mostly used as a helper method for the save method. It returns a string representation of the class. Also fixed a typo in initializeStock() that was reading the price in as an integer instead of a double